Save t state in raycaster
This enables rendering of views where the working set is larger than the
cache. Otherwise at some point in the rendering process, the same 50
bricks will be evicted and loaded again and again. By saving t and
resuming rendering from this point, other bricks (which have been used
for rendering, but are no longer needed for completion) can be evicted
instead.